[nameToFaces-01] fixing loading screen and added the font file
